+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The main work (all changes without a GitHub username in brackets in the below li
 	## __WORK IN PROGRESS__
 -->
 
-## __WORK IN PROGRESS__
+## 0.10.2 (2024-09-15)
 
 -   Matter-Core functionality:
     -   Fix: Fixes channel cleanup